requirements.txt
asgiref==3.3.4
boto3==1.17.84
botocore==1.20.84
colorama==0.4.4
Django==3.2.3
django-crispy-forms==1.11.2
django-storages==1.11.1
gunicorn==20.1.0
jmespath==0.10.0
Pillow==8.2.0
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5